requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=0.19.0
requests>=2.26.0
PyPDF2>=2.0.0
python-docx>=0.8.11
sentence-transformers>=2.2.0
chromadb>=0.3.0
numpy>=1.21.0
pydantic>=1.8.0
pydantic[email]>=1.8.0